api/requirements.txt
# Core Flask dependencies
flask==2.2.2
gunicorn==20.1.0
werkzeug==2.3.7
flask-cors==3.0.10
python-dotenv>=0.21.0,<0.22.0

# Manim and video processing
manim==0.18.0
ffmpeg==1.4

# Cloud storage
azure-storage-blob==12.18.2

# Authentication and database
firebase_admin==6.2.0

# HTTP requests
requests==2.31.0

# AWS services
boto3==1.37.3

# AI/LLM for code safety checks
openai==1.59.9
